Integrate TooltipComponent getter
