+++ packages/wiggle/src/WiggleRendering.scss
@@ -4,19 +4,7 @@
     color: black;
   }
 
-  .hoverVertical {
-    background: #333;
-    border: none;
-    width: 1px;
-    height: 100%;
-    top: 0px;
-    cursor: default;
-    position: absolute;
-  }
-  .hoverLabel {
-    top: 0px;
-    position: absolute;
-  }
+
 
   table,
   th,
